Make spec script for running Specs on Travis
